reduce the workload on UI thread. refs #14
