+++ AnimatedImage.Avalonia/WriteableBitmapFace.cs
@@ -3,9 +3,12 @@ using Avalonia.Controls;
 using Avalonia.Media;
 using Avalonia.Media.Imaging;
 using Avalonia.Platform;
+using Avalonia.Threading;
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace AnimatedImage.Avalonia
 {
@@ -15,83 +18,80 @@ namespace AnimatedImage.Avalonia
         {
             get
             {
-                _flg = !_flg;
                 Send();
-                return Current;
+                return _bitmapWrapper;
             }
         }
 
-        public IImage Current => _flg ? _tick : _tack;
-
-        private bool _flg = false;
         private readonly PixelSize _size;
         private readonly WriteableBitmap _buffer;
         private readonly WriteableBitmap _bitmap;
-        public readonly IImage _tick;
-        public readonly IImage _tack;
+        public readonly IImage _bitmapWrapper;
 
         public WriteableBitmapFace(int width, int height)
         {
             _size = new PixelSize(width, height);
             var dpi = new Vector(96, 96);
             _buffer = new WriteableBitmap(_size, dpi, PixelFormats.Bgra8888, null);
-            _bitmap = new WriteableBitmap(_size, dpi, PixelFormats.Bgra8888, null);
 
-            _tick = new BitmapWrapper(_bitmap);
-            _tack = new BitmapWrapper(_bitmap);
+            _bitmap = new WriteableBitmap(_size, dpi, PixelFormats.Bgra8888, null);
+            _bitmapWrapper = new BitmapWrapper(_bitmap);
         }
 
-        public unsafe void ReadBGRA(byte[] buffer, int x, int y, int width, int height)
+        public void ReadBGRA(byte[] buffer, int x, int y, int width, int height)
         {
             if (width * height * 4 > buffer.Length)
                 throw new IndexOutOfRangeException();
 
             using var bit = _buffer.Lock();
 
-            byte* ptr = (byte*)(void*)(bit.Address);
+            IntPtr ptr = bit.Address;
             ptr += y * bit.RowBytes + x * 4;
 
-            fixed (byte* buffer0 = &buffer[0])
+            var bufferOffset = 0;
+            var copyLen = Math.Min(_size.Width - x, width) * 4;
+
+            for (var i = 0; i < Math.Min(_size.Height - y, height); ++i)
             {
-                byte* bufferPtr = buffer0;
-
-                for (var i = 0; i < Math.Min(_size.Height - y, height); ++i)
-                {
-                    for (var j = 0; j < Math.Min(_size.Width - x, width) * 4; ++j)
-                    {
-                        bufferPtr[j] = ptr[j];
-                    }
-
-                    bufferPtr += width * 4;
-                    ptr += bit.RowBytes;
-                }
+                Marshal.Copy(ptr, buffer, bufferOffset, copyLen);
+                bufferOffset += width * 4;
+                ptr += bit.RowBytes;
             }
         }
 
-        public unsafe void WriteBGRA(byte[] buffer, int x, int y, int width, int height)
+        public void WriteBGRA(byte[] buffer, int x, int y, int width, int height)
         {
             if (width * height * 4 > buffer.Length)
                 throw new IndexOutOfRangeException();
 
             using var bit = _buffer.Lock();
 
-            byte* ptr = (byte*)(void*)(bit.Address);
+            IntPtr ptr = bit.Address;
             ptr += y * bit.RowBytes + x * 4;
 
-            fixed (byte* buffer0 = &buffer[0])
+            var bufferOffset = 0;
+            var copyLen = Math.Min(_size.Width - x, width) * 4;
+
+            for (var i = 0; i < Math.Min(_size.Height - y, height); ++i)
             {
-                byte* bufferPtr = buffer0;
+                Marshal.Copy(buffer, bufferOffset, ptr, copyLen);
+                bufferOffset += width * 4;
+                ptr += bit.RowBytes;
+            }
+        }
 
-                for (var i = 0; i < Math.Min(_size.Height - y, height); ++i)
-                {
-                    for (var j = 0; j < Math.Min(_size.Width - x, width) * 4; ++j)
-                    {
-                        ptr[j] = bufferPtr[j];
-                    }
-                    bufferPtr += width * 4;
+        public unsafe void Clear(int x, int y, int width, int height)
+        {
+            using var bit = _buffer.Lock();
 
-                    ptr += bit.RowBytes;
-                }
+            IntPtr ptr = bit.Address;
+            ptr += y * bit.RowBytes + x * 4;
+
+            var copyLen = (uint)Math.Min(_size.Width - x, width) * 4;
+            for (var i = 0; i < Math.Min(_size.Height - y, height); ++i)
+            {
+                Unsafe.InitBlock(ptr.ToPointer(), 0, copyLen);
+                ptr += bit.RowBytes;
             }
         }
