+++ AnimatedImage/FrameRenderer.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
+using System.Threading.Tasks;
 using AnimatedImage.Formats;
 using AnimatedImage.Formats.Gif;
 using AnimatedImage.Formats.Png;
@@ -75,6 +76,8 @@ namespace AnimatedImage
         /// frameindex is less than 0, or larger than or equals to the count of frames.
         /// </exception>
         public abstract void ProcessFrame(int frameIndex);
+
+        public abstract Task ProcessFrameAsync(int frameIndex);
 
         /// <summary>
         /// Retrieves the elapsed time until the specified frame is drawn. If the first frame is specified, returns 0 seconds.
@@ -166,70 +169,6 @@ namespace AnimatedImage
 
                 return true;
             }
-        }
-    }
-
-    /// <summary>
-    /// Encapsulates frame area and drawing time.
-    /// </summary>
-    public class FrameRenderFrame
-    {
-        /// <summary>
-        /// The left positions of the frame area.
-        /// </summary>
-        public int X { get; }
-        /// <summary>
-        /// The top positions of the frame area.
-        /// </summary>
-        public int Y { get; }
-        /// <summary>
-        /// The width of the frame area.
-        /// </summary>
-        public int Width { get; }
-        /// <summary>
-        /// The height of the frame area.
-        /// </summary>
-        public int Height { get; }
-
-        /// <summary>
-        /// The elapsed time until the specified frame is drawn
-        /// </summary>
-        public TimeSpan Begin { get; }
-        /// <summary>
-        /// The elapsed time until the drawing result of the specified frame is disposes
-        /// </summary>
-        public TimeSpan End { get; }
-
-        /// <summary>
-        /// Creates the insntace.
-        /// </summary>
-        /// <param name="x"><see cref="X"/></param>
-        /// <param name="y"><see cref="Y"/></param>
-        /// <param name="width"><see cref="Width"/></param>
-        /// <param name="height"><see cref="Height"/></param>
-        /// <param name="begin"><see cref="Begin"/></param>
-        /// <param name="end"><see cref="End"/></param>
-        public FrameRenderFrame(int x, int y, int width, int height, TimeSpan begin, TimeSpan end)
-        {
-            X = x;
-            Y = y;
-            Width = width;
-            Height = height;
-            Begin = begin;
-            End = end;
-        }
-
-        /// <summary>
-        /// Determines whether this instance contains the specified area.
-        /// </summary>
-        /// <param name="frame">The specified area.</param>
-        /// <returns>Returns true if this instance contains the specified area.</returns>
-        public bool IsInvolve(FrameRenderFrame frame)
-        {
-            return X <= frame.X
-                && Y <= frame.Y
-                && frame.X + frame.Width <= X + Width
-                && frame.Y + frame.Height <= Y + Height;
         }
     }
 }
